% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpb.R
\name{createPB}
\alias{createPB}
\title{Create a progress bar for calling from fortran}
\usage{
createPB(...)
}
\arguments{
\item{...}{args to \code{utils::txtProgressBar} including min and max}
}
\description{
Create a progress bar for calling from fortran
}
