% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{detect_long_lines}
\alias{detect_long_lines}
\title{Detect if there are long lines in mortran or fortran sections}
\usage{
detect_long_lines(mortran_lines)
}
\arguments{
\item{mortran_lines}{the mortran file lines resulting from a
\code{base::readLines()}, say.}
}
\value{
a possibly empty data frame of the approximate line number
and the offending line if any
}
\description{
Detect if there are long lines in mortran or fortran sections
}
\examples{
\dontrun{
check_long_lines(readLines("./pcLasso.m"))
}

}
