% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{fix_unused_labels}
\alias{fix_unused_labels}
\title{Fix unused label warnings from gfortran if possible}
\usage{
fix_unused_labels(fortran_lines, verbose = FALSE)
}
\arguments{
\item{verbose}{a flag for verbose output}

\item{mortran_lines}{the mortran file lines resulting from a
\code{base::readLines()}, say.}
}
\value{
fortran lines as a character vector
}
\description{
Fix unused label warnings from gfortran if possible
}
\examples{
\dontrun{
fix_unused_labels(readLines("foo.f"))
}

}
