% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{fortran_subroutines}
\alias{fortran_subroutines}
\title{Return all fortran subroutines with args}
\usage{
fortran_subroutines(lines)
}
\arguments{
\item{lines}{character vector of lines of a mortran file}
}
\value{
a list of subroutines
}
\description{
Return all fortran subroutines with args
}
\examples{
\dontrun{
fortran_subroutines("pclasso.m")
}
}
