% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{generate_fortran}
\alias{generate_fortran}
\title{Generate fortran from a mortran file and return the fortran}
\usage{
generate_fortran(mortran_lines)
}
\arguments{
\item{mortran_lines}{the mortran file lines resulting from a
\code{base::readLines()}, say.}
}
\value{
fortran lines as a character vector
}
\description{
Generate fortran from a mortran file and return the fortran
}
\examples{
\dontrun{
generate_fortran(readLines("./pcLasso.m"))
}

}
