% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{make_dotcall_to_fortran}
\alias{make_dotcall_to_fortran}
\title{Generate C stub and R stub for making a .Call to a Fortran function}
\usage{
make_dotcall_to_fortran(fun_strs)
}
\arguments{
\item{fun_str}{a vector of subroutine definition strings with args}
}
\value{
a named list of C code and R code for each function
}
\description{
Parse a fortran subroutine/entry definition and return a C routine that can be used with .Call and R call snippet
}
\examples{
make_dotcall_to_fortran(c("subroutine foo(no,ni,x,y,w,theta)", "entry foobar(ix, y)", "subroutine bar(x,n,p,m)"))

}
