% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{make_fun_list}
\alias{make_fun_list}
\title{Generate function name and C arg types from subroutine/entry definition}
\usage{
make_fun_list(fun_strs)
}
\arguments{
\item{fun_str}{a vector of subroutine definition strings with args}
}
\value{
a list of function names along with C argument types, SEXP args, and .Call args
}
\description{
Parse a fortran subroutine/entry definition and return subroutine/entry name and
C argument types as a list
}
\examples{
make_fun_list(c("subroutine foo(no,ni,x,y,w,theta)", "entry foobar(ix, y)", "subroutine bar(x,n,p,m)"))


}
