% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{make_sexp_args}
\alias{make_sexp_args}
\title{Make SEXP argument types}
\usage{
make_sexp_args(...)
}
\arguments{
\item{...}{list of variable names, as is}
}
\description{
Make SEXP argument types
}
\examples{
make_sexp_args(no,ni,x,y,w,theta,ng,mg,aa,ne,nx,nlam,ulam,thr,maxit,verbose,ao,ia,kin,nlp,jerr)


}
