% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{mortran_subroutines}
\alias{mortran_subroutines}
\title{Return all mortran subroutines with args}
\usage{
mortran_subroutines(lines)
}
\arguments{
\item{lines}{character vector of lines of a mortran file}
}
\value{
a list of subroutines
}
\description{
Return all mortran subroutines with args
}
\examples{
\dontrun{
mortran_subroutines("pclasso.m")
}
}
