% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_mortran}
\alias{process_mortran}
\title{Generate fortran and registration code from mortran, geared
specifically towards the mortran used by Jerome Friedman at
Stanford University}
\usage{
process_mortran(input_mortran_file, pkg_name = NULL, verbose = TRUE)
}
\arguments{
\item{input_mortran_file}{the input mortran file}

\item{pkg_name}{the package name signalling that registration code
has to be generated}

\item{verbose}{a flag for verbose output}
}
\value{
a named list of processed mortran lines (\code{mortran}),
fortran lines (\code{fortran}) and, if so requested and no warnings
occur, a registration c code as the third argument
appropriately named using the package name provided
}
\description{
Generate fortran and registration code from mortran, geared
specifically towards the mortran used by Jerome Friedman at
Stanford University
}
\examples{
\dontrun{
process_mortran("./pcLasso.m")
}

}
