% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUtools-package.R
\docType{package}
\name{SUtools-package}
\alias{SUtools-package}
\alias{cubature-package}
\alias{cubature}
\title{SUtools is a package for Fortran code handling with R}
\description{
Tools for autogenerating fortran registration code. I must have done
this at least three times and forgot to organize the code each time.
This package bundles the \code{mortran} preprocessor, with slight
modifications.
}
\author{
Balasubramanian Narasimhan
Maintainer: Balasubramanian Narasimhan\href{mailto:naras@stat.stanford.edu}{naras@stat.stanford.edu}
}
\keyword{package}
