% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{sutools_control}
\alias{sutools_control}
\title{SUtools Control Arguments}
\usage{
sutools_control(fix_allocate = FALSE, verbose = TRUE)
}
\arguments{
\item{fix_allocate}{a flag denoting whether Jerome Friedman's allocate statements have to be fixed, default \code{FALSE}}

\item{verbose}{a flag asking for progress notes, default \code{TRUE}}
}
\value{
a list containing the control parameters.
}
\description{
Details to the \emph{control} parameters.
}
