% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_mortran}
\alias{process_mortran}
\title{Generate fortran and registration code from mortran}
\usage{
process_mortran(input_mortran_file, register = FALSE, pkg_name = NULL,
  output_fortran_file = paste0(tools::file_path_sans_ext(basename(input_mortran_file)),
  ".f"))
}
\arguments{
\item{input_mortran_file}{the input mortran file}

\item{register}{if TRUE, will generate a registration pkg_init.c file automatically and pkg_name must be supplied}

\item{pkg_name}{see register argument above, must be the actual package name}

\item{output_mortran_file}{the output fortran file, optional}
}
\description{
Generate fortran and registration code from mortran
}
\examples{
\dontrun{
process_file("./pcLasso.m")
}

}
